requirements.txt
pandas==2.2.3
python_binance==1.0.19
streamlit==1.41.1
